Moved ghost_protocol.rs into dht/dht_protocol.rs
